metrics_analyzer/requirements.txt
psycopg2-binary
pandas
matplotlib
sqlalchemy